created compounds subpackage
