+++ prody/compounds/__init__.py
@@ -1,0 +1,25 @@
+# -*- coding: utf-8 -*-
+"""This module defines classes and functions to fetch, parse, and write
+structural data files associated with compounds and to access and search 
+structural databases, e.g. `ProteinDataBank <http://wwpdb.org>`_.
+
+Ligand data
+===========
+
+Following function can be used to fetch meta data on PDB ligands:
+
+  * :func:`.fetchPDBLigand` - retrieve ligand from Ligand-Expo
+
+"""
+
+__all__ = []
+
+from . import pdbligands
+from .pdbligands import *
+__all__.extend(pdbligands.__all__)
+
+from . import functions
+from .functions import *
+__all__.extend(functions.__all__)
+
+
